AI_Scholar_v1/requirements.txt
transformers
langchain
faiss-cpu
torch
dotenv
